# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Hamlaneh API
  description: >
    Self-hosted team communication platform API. Phase 1.1 surface: health,
    authentication, current user, and the admin user-management endpoints that
    back the dashboard (public registration is off by default; users are
    created from the admin dashboard or via invite links). Phase 1.2 adds the
    messaging surface the chat shell needs: channels and 1:1 direct messages
    with membership, cursor-paged message history with idempotent send, edit
    and soft delete, per-user read positions, message search, and the
    WebSocket upgrade that carries realtime delivery. Phase 1.1b completes the
    identity surface: email password reset, TOTP two-step verification with
    recovery codes, and self-service session management.
  version: 0.6.0
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /
    description: Same-origin deployment behind Caddy.
...
paths:
...
  /api/v1/channels/{channelId}/call:
    get:
      tags: [channels]
      operationId: getChannelCall
      summary: What is happening in this channel's call.
      description: >
        Live state, read from the media server rather than from a table.
        There is no calls table: the media server's own room state is the
        truth, and a copy in the database would be a cache to invalidate for
        nothing (ADR 005).

        Clients call this on opening a channel and after a reconnect. The
        three call events on the WebSocket are hints that something changed,
        never the state itself -- a five-minute-old call event is worthless
        or wrong, so it carries no sequence number and is never replayed.
      security:
        - sessionCookie: []
      parameters:
        - name: channelId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: The call, or that there is not one.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelCall"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/channels/{channelId}/call/token:
    post:
      tags: [channels]
      operationId: createCallToken
      summary: A ticket to join this channel's call.
      description: >
        Mints a short-lived join ticket for the media server. It is scoped to
        one room and one identity and lives about two minutes, because a
        ticket has no business outliving the click that asked for it.

        It is not a session and grants nothing else: it cannot enumerate
        rooms, cannot eject anyone, and cannot open a data channel -- chat
        stays on the one write path with the one authorization choke point.

        A ticket can outlive the membership that justified it, briefly, and
        that is bounded rather than denied. Minted-but-unused is bounded by
        the expiry. Already-joined is bounded by the server ejecting the
        participant when membership or the account ends. Both windows and
        their residue are written out in ADR 005.

        The media server's address is not in the answer: its signal endpoint
        is same-origin, so a client derives it the way it derives everything
        else. Nothing about the media plane is ever told to a client beyond
        one boolean on the instance document.
      security:
        - sessionCookie: []
      parameters:
        - name: channelId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "201":
          description: The ticket.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallToken"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          description: >
            Calls are not configured on this instance (code calls_unavailable).
            Chat is unaffected.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: hamlaneh_session
      description: >
        HttpOnly+Secure+SameSite=Strict session cookie. Mutating requests must
        also send the X-Hamlaneh-CSRF header.
...
  responses:
...
    Unauthorized:
      description: Missing, expired, or invalid session/credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: >
        No such resource for this caller. Channel-scoped paths answer 404 —
        never 403 — to a non-member, so a channel's existence never leaks; an
        org admin who is not a member sees the same 404.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: >
        Too many attempts; retry later. Always carries Retry-After, so a client
        can show a real countdown instead of guessing — without it a form has no
        honest way to say when the door reopens.
      headers:
        Retry-After:
          description: >
            Whole seconds until the caller's budget frees up. Present on every
            429 this server sends: one function writes them all and it sets the
            header unconditionally, so carrying it is a property of the code
            rather than something each handler has to remember.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable code (e.g. invalid_credentials).
            message:
              type: string
              description: Human-readable English message; clients localize by code.
...
    Presence:
      type: string
      description: >
        online — a live socket. away — the client reported itself idle.
        offline — no socket after a short grace period.
      enum: [online, away, offline]
    UserSummary:
      type: object
      description: >
        The public face of a user: everything the chat shell draws (name row,
        avatar initials and tint are derived client-side from display_name and
        id) and nothing else. Never carries email, role, or password state.
      required: [id, username, display_name]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        display_name:
          type: string
        presence:
          allOf:
            - $ref: "#/components/schemas/Presence"
          description: >
            Present only where the design shows presence — the DM peer and the
            user directory. Message authors carry no presence.
...
    CallParticipant:
      type: object
      required: [user, joined_at]
      properties:
        user: { $ref: "#/components/schemas/UserSummary" }
        joined_at: { type: string, format: date-time }
        screen_sharing:
          type: boolean
          description: Whether this participant is publishing a screen share.
    ChannelCall:
      type: object
      description: >
        A channel's live call. `active` false means nobody is in it, and the
        other fields are absent rather than stale.
      required: [active]
      properties:
        active: { type: boolean }
        started_at: { type: string, format: date-time }
        participants:
          type: array
          items:
            $ref: "#/components/schemas/CallParticipant"
    CallToken:
      type: object
      required: [token, room, expires_at]
      properties:
        token:
          type: string
          description: >
            The join ticket, for the media client. Short-lived, single room,
            single identity. Never store it; ask again on the next join.
        room:
          type: string
          description: >
            The room to join. Derived from the channel, and safe to hand back
            because no ticket this server mints can enumerate rooms -- the
            only people who ever see this name are already members of the
            channel it names.
        expires_at: { type: string, format: date-time }
...